Make ChannelActionListener a top-level class (#37797)
We start using this class more often. Let's make it a top-level class.
