+++ server/src/main/java/org/elasticsearch/action/support/ChannelActionListener.java
@@ -1,0 +1,62 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.elasticsearch.action.support;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.message.ParameterizedMessage;
+import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.transport.TransportChannel;
+import org.elasticsearch.transport.TransportRequest;
+import org.elasticsearch.transport.TransportResponse;
+
+public final class ChannelActionListener<
+    Response extends TransportResponse, Request extends TransportRequest> implements ActionListener<Response> {
+
+    private static final Logger logger = LogManager.getLogger(ChannelActionListener.class);
+    private final TransportChannel channel;
+    private final Request request;
+    private final String actionName;
+
+    public ChannelActionListener(TransportChannel channel, String actionName, Request request) {
+        this.channel = channel;
+        this.request = request;
+        this.actionName = actionName;
+    }
+
+    @Override
+    public void onResponse(Response response) {
+        try {
+            channel.sendResponse(response);
+        } catch (Exception e) {
+            onFailure(e);
+        }
+    }
+
+    @Override
+    public void onFailure(Exception e) {
+        try {
+            channel.sendResponse(e);
+        } catch (Exception e1) {
+            logger.warn(() -> new ParameterizedMessage(
+                "Failed to send error response for action [{}] and request [{}]", actionName, request), e1);
+        }
+    }
+}
